Fixed issues with retrieving task lists with one task
When retrieving a task list, the JSON format of ISPW's response is inconsistent if the list has only one entry in it -- the entry is then returned as a JSON structure, not as a list containing one JSON structure. Added code to compensate for this so that our response format is consistent. Also fixed various descriptions and other minor issues.